Create a filter score object

score_vec() creates a score_vec object from a double vector. The direction of best result (e.g. maximization) can be set along with a numeric value to use when the value cannot be computed.

score_vec(
  x = numeric(),
  direction = "maximize",
  impute = Inf,
  call = rlang::caller_env()
)

Arguments

x

A numeric vector..

direction

What is the best result? Possible values are "maximize", "minimize", or "zero".

impute

A single numeric value for when values are missing. This should be a value that indicates that a variable should not be removed.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of rlang::abort() for more information.

Details

direction values are those that you feel unsure about, and would like to exclude from performance calculations or other metrics.

Examples

set.seed(1)
score_vec(rnorm(10), "zero")
#>  [1] -0.626  0.184 -0.836  1.595  0.330 -0.820  0.487  0.738  0.576 -0.305
#> Direction: zero